# backend/docs/swagger.yaml
definitions:
  gin.H:
    additionalProperties: {}
    type: object
  model.Image:
    properties:
      height:
        description: The image height, in pixels.
        type: integer
      url:
        description: The source URL of the image.
        type: string
      width:
        description: The image width, in pixels.
        type: integer
    type: object
  model.InclusionResponse:
    properties:
      included:
        $ref: '#/definitions/model.InclusionType'
      spotifyID:
        type: string
    type: object
  model.InclusionType:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    type: integer
    x-enum-varnames:
    - Nothing
    - Included
    - Excluded
    - IncludedByProxy
    - ExcludedByProxy
  model.ItemInclusionRequest:
    properties:
      include:
        type: boolean
      playlistid:
        type: string
      spotid:
        type: string
      type:
        $ref: '#/definitions/model.ItemType'
    required:
    - include
    - playlistid
    - spotid
    - type
    type: object
  model.ItemPlaylistRequest:
    properties:
      cspotid:
        type: string
      pspotid:
        type: string
    required:
    - cspotid
    - pspotid
    type: object
  model.ItemRequest:
    properties:
      parentid:
        type: string
      playlistid:
        type: string
      type:
        $ref: '#/definitions/model.ItemType'
    required:
    - parentid
    - playlistid
    - type
    type: object
  model.ItemResponse:
    properties:
      icon:
        items:
          $ref: '#/definitions/model.Image'
        type: array
      included:
        $ref: '#/definitions/model.InclusionType'
      inclusionByProxy:
        type: boolean
      itemType:
        $ref: '#/definitions/model.ItemType'
      name:
        type: string
      sortdata:
        type: integer
      spotifyID:
        type: string
    required:
    - included
    - name
    - sortdata
    - spotifyID
    type: object
  model.ItemType:
    enum:
    - 0
    - 1
    - 2
    - 3
    type: integer
    x-enum-varnames:
    - PlaylistItem
    - ArtistItem
    - AlbumItem
    - TrackItem
  model.PlaylistCreateRequest:
    properties:
      name:
        example: My Playlist
        type: string
    required:
    - name
    type: object
  model.PlaylistPublishRequest:
    properties:
      spotifyID:
        type: string
    type: object
  model.PlaylistResponse:
    properties:
      name:
        type: string
      spotifyID:
        type: string
    type: object
  model.SearchRequest:
    properties:
      playlistid:
        type: string
      query:
        type: string
      type:
        $ref: '#/definitions/model.ItemType'
    required:
    - playlistid
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /auth/status:
    get:
      description: Returns true if the backend has a valid Spotify connection.
      produces:
      - application/json
      responses:
        "200":
          description: 'authenticated: true'
          schema:
            additionalProperties:
              type: boolean
            type: object
      summary: Check authentication status
      tags:
      - auth
  /auth/url:
    get:
      description: Returns the URL the user needs to visit to authorize the app on
        Spotify.
      produces:
      - application/json
      responses:
        "200":
          description: 'url: https://accounts.spotify.com/...'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Spotify Authorization URL
      tags:
      - auth
  /playlist:
    delete:
      description: Deletes every playlist record in the database
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Clear all playlists
      tags:
      - playlist
    get:
      description: Responds with the list of all playlists as JSON.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.PlaylistResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all playlists
      tags:
      - playlist
    post:
      consumes:
      - application/json
      description: Create a new playlist locally and on Spotify
      parameters:
      - description: Playlist name
        in: body
        name: playlist
        required: true
        schema:
          $ref: '#/definitions/model.PlaylistCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.PlaylistResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create new playlist
      tags:
      - playlist
  /playlist/{id}:
    delete:
      description: Delete a specific playlist by its ID
      parameters:
      - description: Playlist ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a playlist
      tags:
      - playlist
  /playlist/{id}/exclusions:
    get:
      consumes:
      - application/json
      description: Fetches the list of all Artists, Albums, and Tracks manually excluded
        in a specific playlist.
      parameters:
      - description: Spotify Playlist ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ItemResponse'
            type: array
        "400":
          description: 'error: Invalid ID'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Database error'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all excluded items for a playlist
      tags:
      - playlist
  /playlist/{id}/inclusions:
    get:
      consumes:
      - application/json
      description: Fetches the list of all Playlists, Artists, Albums, and Tracks
        manually included in a specific playlist.
      parameters:
      - description: Spotify Playlist ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ItemResponse'
            type: array
        "400":
          description: 'error: Invalid ID'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Database error'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all included items for a playlist
      tags:
      - playlist
  /playlist/{id}/playlists:
    get:
      description: Responds with the list of all playlists as JSON.
      parameters:
      - description: Playlist ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ItemResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all playlists as itemresponse items
      tags:
      - playlist
  /playlist/{id}/rename:
    put:
      consumes:
      - application/json
      description: Updates the display name of a playlist in the local database.
      parameters:
      - description: Spotify Playlist ID
        in: path
        name: id
        required: true
        type: string
      - description: New name for the playlist
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.PlaylistCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Success'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: Invalid input'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Playlist not found'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Database error'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Rename a playlist
      tags:
      - playlist
  /playlist/include:
    post:
      consumes:
      - application/json
      description: Includes one playlist inside another parent playlist
      parameters:
      - description: Playlist Linking Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.ItemPlaylistRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PlaylistResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Nest a Playlist
      tags:
      - playlist
  /playlist/include/undo:
    post:
      consumes:
      - application/json
      description: Undoes inclusion of one playlist inside another parent playlist
      parameters:
      - description: Playlist Linking Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.ItemPlaylistRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PlaylistResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Undo a nested Playlist
      tags:
      - playlist
  /playlist/item:
    post:
      consumes:
      - application/json
      description: Adds an IdItem to either the inclusions or exclusions of a playlist
      parameters:
      - description: Inclusion/Exclusion Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.ItemInclusionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.InclusionResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Include or Exclude an Item
      tags:
      - items
  /playlist/item/undo:
    post:
      consumes:
      - application/json
      description: Removes an IdItem from both inclusions and exclusions of a playlist
      parameters:
      - description: Item to remove
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.ItemInclusionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.InclusionResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Undo Include or Exclude
      tags:
      - items
  /playlist/publish:
    post:
      consumes:
      - application/json
      description: Calculates the current tracklist based on inclusions/exclusions
        and replaces the Spotify playlist content.
      parameters:
      - description: Playlist Publish Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.PlaylistPublishRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Success'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: Bad Request'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Internal Server Error'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Publish a playlist to Spotify
      tags:
      - playlist
  /playlist/publishall:
    post:
      consumes:
      - application/json
      description: Calculates the current tracklist based on inclusions/exclusions
        and replaces the Spotify playlist content.
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Success'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: Bad Request'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Publish all playlists to Spotify
      tags:
      - playlist
  /search:
    post:
      consumes:
      - application/json
      description: Search Spotify for Artists, Albums, or Tracks based on the provided
        ItemType.
      parameters:
      - description: Search Query and Item Type
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.SearchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ItemResponse'
            type: array
        "400":
          description: Invalid Item Type
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Unified Spotify Search
      tags:
      - search
  /spotify/album/tracks:
    post:
      consumes:
      - application/json
      description: Fetches all tracks for a specific Spotify album
      parameters:
      - description: Album and Playlist Context
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.ItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ItemResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Tracks by Album
      tags:
      - spotify
  /spotify/artist/albums:
    post:
      consumes:
      - application/json
      description: Fetches albums and singles from Spotify for a specific artist
      parameters:
      - description: Artist and Playlist Context
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.ItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.ItemResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Albums by Artist
      tags:
      - spotify
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"